User Request Router Rough In

+++ src/routers/groupRouters/groupIDRouters/groupUserRouter.ts
@@ -61,15 +61,18 @@ groupUserRouter.post("/request",
 
                     switch (req.body.action) {
                         case "accept_request":
-                            userIDs = await GroupModel.retrieve_user_id_by_username(reqValues.usernames, reqValues.groupID, "user_request_active")
+                            userIDs = await GroupModel.retrieve_user_id_by_username(reqValues.usernames, reqValues.groupID, "user_request_active");
+                            if (userIDs.length < 1) {throw new ExpressError("Server Error: Unable to add users.", 500);};
                             queryData = await GroupModel.create_group_user(reqValues.groupID, userIDs);
-                            return res.json({reqAccept: queryData})
+                            return res.json({reqAccept: queryData});
                         case "send_request":
-                            userIDs = await GroupModel.retrieve_user_id_by_username(reqValues.usernames, reqValues.groupID, "invite_permitted")
+                            userIDs = await GroupModel.retrieve_user_id_by_username(reqValues.usernames, reqValues.groupID, "group_request_permitted");
+                            if (userIDs.length < 1) {throw new ExpressError("Server Error: Unable to invite user to join group.", 500);};
                             queryData = await GroupModel.create_request_group_to_user(reqValues.groupID, userIDs);
-                            return res.json({reqSent: queryData})
+                            return res.json({reqSent: queryData});
                         case "remove_request":
-                            userIDs = await GroupModel.retrieve_user_id_by_username(reqValues.usernames)
+                            userIDs = await GroupModel.retrieve_user_id_by_username(reqValues.usernames, reqValues.groupID);
+                            if (userIDs.length < 1) {throw new ExpressError("Server Error: Unable to remove invite request.", 500);};
                             queryData = await GroupModel.delete_request_user_group(userIDs, reqValues.groupID);
                             return res.json({reqRemove: queryData})
                         default:
